+++ src/screens/Home/catalogs/styles.ts
